feat(freezeV2): optimize calculation and add unit tests for freezeV2 precompiledContracts

+++ actuator/src/main/java/org/tron/core/vm/utils/FreezeV2Util.java
@@ -152,7 +152,7 @@ public class FreezeV2Util {
         return 0L;
       }
 
-      return frozenV2Resource * (totalResource - usage) / totalResource;
+      return (long) (frozenV2Resource * ((double) (totalResource - usage) / totalResource));
     }
 
     if (type == 1) {
@@ -177,7 +177,7 @@ public class FreezeV2Util {
         return 0L;
       }
 
-      return frozenV2Resource * (totalResource - usage) / totalResource;
+      return (long) (frozenV2Resource * ((double) (totalResource - usage) / totalResource));
     }
 
     return 0L;
